# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BullBearBroker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/push/subscribe": {
      "post": {
        "tags": [
          "push",
          "push"
        ],
        "summary": "Subscribe Push",
        "operationId": "subscribe_push_api_push_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/push/preferences": {
      "put": {
        "tags": [
          "push",
          "push"
        ],
        "summary": "Update Preferences",
        "operationId": "update_preferences_api_push_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushPreferencesPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushPreferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/push/send-test": {
      "post": {
        "tags": [
          "push",
          "push"
        ],
        "summary": "Send Test Notification",
        "operationId": "send_test_notification_api_push_send_test_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Test Notification Api Push Send Test Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushPreferencesPayload": {
        "properties": {
          "alerts": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerts"
          },
          "news": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "News"
          },
          "system": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          }
        },
        "type": "object",
        "title": "PushPreferencesPayload",
        "example": {
          "alerts": true,
          "news": false,
          "system": true
        }
      },
      "PushPreferencesResponse": {
        "properties": {
          "alerts": {
            "type": "boolean",
            "title": "Alerts"
          },
          "news": {
            "type": "boolean",
            "title": "News"
          },
          "system": {
            "type": "boolean",
            "title": "System"
          }
        },
        "type": "object",
        "required": [
          "alerts",
          "news",
          "system"
        ],
        "title": "PushPreferencesResponse",
        "example": {
          "alerts": true,
          "news": true,
          "system": true
        }
      },
      "PushSubscriptionPayload": {
        "properties": {
          "endpoint": {
            "type": "string",
            "minLength": 10,
            "title": "Endpoint"
          },
          "keys": {
            "$ref": "#/components/schemas/SubscriptionKeys"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscriptionPayload",
        "example": {
          "endpoint": "https://fcm.googleapis.com/fcm/send/abc123",
          "keys": {
            "auth": "auth-secret",
            "p256dh": "client-public-key"
          }
        }
      },
      "PushSubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PushSubscriptionResponse"
      },
...
      "SubscriptionKeys": {
        "properties": {
          "auth": {
            "type": "string",
            "minLength": 4,
            "title": "Auth"
          },
          "p256dh": {
            "type": "string",
            "minLength": 4,
            "title": "P256Dh"
          }
        },
        "type": "object",
        "required": [
          "auth",
          "p256dh"
        ],
        "title": "SubscriptionKeys"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}